requirements.txt
Pillow>=10.1.0
easyocr>=1.7.0
torch>=2.0.0
torchvision>=0.15.0
